+++ lib/provider/provider_set_up.dart
@@ -15,7 +15,7 @@ import '../data/repositories/taleant_board_repository.dart';
 import '../main.dart';
 import '../utils/token_manager.dart';
 import '../view_model/auth_view_model.dart';
-import '../view_model/taleant_board_view_model.dart';
+import '../view_model/talent_board_view_model.dart';
 import 'auth/kakao_provider.dart';
 import 'common/common_provider.dart';
